refactor: move chunk config resolver
